Fix user removed from local groups
Only remove the user from local groups which the user was added to by
authd. Previously, authd removed the user from all local groups which
are not configured in Microsoft Entra, so if the user was added to any
local groups manually, it was removed from those groups again during the
next login.

Closes #576